Attempting to solve 1: Two Sum
Trying a loop within a loop.

For each input returns 2 numbers which sum to the target.

BUT it:
1) returns numbers both ways around e.g. 1 + 2 and 2 + 1 when the first result is sufficient.
2) reuses a number at the same index twice e.g. to get 6 it will use 3 at index 0 twice, instead of another 2 numbers e.g. 4 and 2.

Next I want to instead build a table with each possible pair, its constituent's original indices and their sum. Then lookup the target and return the corresponding indices from the table. HASH MAP / DICTIONARY...